Added cashback logic and test case

+++ customer-service/src/main/java/com/infosys/model/Customer.java
@@ -19,5 +19,5 @@ public class Customer {
 	private String customerName;
 	private String userName;
 	private List<ProductsBought> productsBought;
-	private long cashBackPoints;
+	private double cashBackPoints;
 }
